minimal working tensor bp api

+++ src/graph/tensor_initializers.py
@@ -1,4 +1,4 @@
-from typing import Callable, List
+from typing import Callable
 import jax.numpy as jnp
 from jax.random import normal
 from jax import Array
@@ -7,10 +7,10 @@ from .node import Node
 
 def _gen_ghz_core(phys_dimension: int, modes_number: int) -> Array:
     if phys_dimension == 1:
-        return jnp.array(1.0, dtype=jnp.complex64).reshape(modes_number * (1,))
+        return jnp.array(1.0, dtype=jnp.complex128).reshape(modes_number * (1,))
     stride = (phys_dimension**modes_number - 1) // (phys_dimension - 1)
     elements_number = phys_dimension**modes_number
-    core = jnp.zeros(elements_number, dtype=jnp.complex64)
+    core = jnp.zeros(elements_number, dtype=jnp.complex128)
     core = core.at[0:elements_number:stride].set(1.0)
     return core.reshape(modes_number * (phys_dimension,))
 
@@ -67,7 +67,7 @@ def get_tensor_bloated_ghz_initializer(key: Array) -> Callable[[Node], Array]:
                     f"For GHZ initializer all bond indices must be >= the physical index, got one of the bond dimensions {bond_dim} and physical dimension {node.dimension}."
                 )
             # Key is fixed for the purpose: for fixed shape one have to produce the same matrix
-            bloater = normal(key, (bond_dim, node.dimension)).astype(jnp.complex64)
+            bloater = normal(key, (bond_dim, node.dimension)).astype(jnp.complex128)
             bloater, _ = jnp.linalg.qr(bloater)
             ghz_core = jnp.tensordot(bloater, ghz_core, axes=[1, -1])
         assert ghz_core.shape == (
